requirements.txt
flask
pyyaml
mysql-connector-python
